yamcs-web. Moved SelectParameterDialog to shared module as it can be used for MDB and Monitor
