+++ yamcs-web/packages/app/src/app/shared/dialogs/SelectParameterDialog.html
